requirements.txt
asgiref==3.8.1
Django==4.2.7
django-ckeditor==6.7.2
django-ckeditor-5==0.2.15
django-ckeditor5==0.0.4
django-cors-headers==4.6.0
django-cuser==2017.3.16
django-jazzmin==2.6.0
django-js-asset==3.0.1
django-modeltranslation==0.19.12
django-tinymce==4.1.0
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
modeltranslation==0.25
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
